requirements.txt
```bash
    pip install -r requirements.txt
```
flask~=3.0.0
Flask-Login~=0.7.0
flask-socketio~=5.3.6
werkzeug~=3.0.1
pymongo~=4.5.0
python-dotenv~=1.0.0
openai~=1.0.0rc1